+++ src/commands/build.rs
@@ -174,6 +174,7 @@ pub fn run(args: &BuildArgs) -> anyhow::Result<()> {
     let kernel = kernel_cache::ensure_kernel(
         false,
         args.kernel_config_fragment.clone(),
+        args.module_signing_cert.clone(),
         args.kernel_builder_package.clone(),
     )?;
     println!(
@@ -499,6 +500,11 @@ pub fn run(args: &BuildArgs) -> anyhow::Result<()> {
                     .kernel_extra_config_sha256
                     .clone(),
                 snapshot_config_sha256: kernel.manifest.inputs.snapshot_config_sha256.clone(),
+                module_signing_cert_sha256: kernel
+                    .manifest
+                    .inputs
+                    .module_signing_cert_sha256
+                    .clone(),
             }),
             initrd: manifest::FileEntry {
                 path: manifest::basename_of(&initrd_path),